Add possibility to connect ledger ethereum account
As of now it will only allow to use goerli accounts. Left TODOs around the
places to not forget to fix that.

I've created `useRequestEthereumAccount` which actually works similar to
`useRequestAccount` from `@ledgerhq/wallet-api-client-react`, but it uses our
ledgerLiveManager under the hood. This way it will save the account in our
signer, which then will be used in our tbtc-v2 SDK to interact with contracts.

Since we save our connected address in redux store we will now display
it when the address is connected instead of getting the `account`
property from `useWeb3React` hook.