Make it possible to connect bitcoin account
We want to connect bitcoin account at step 2 of the minting flow. For this case
I've created a separate button that allows the user to connect a bitcoin account
inside ledger live app. After that he will be able to send the bitcoins form the
account he had choosen to a deposit address that he just generated.