Rewamp (1) - install wallet-connect-client-react
I've decided to rewamp the whole implementation as it seems that we don't really
need to keep everything in ledgerLiveAppManager. The idea now is to keep only
LedgerLiveAppEthereumSigner (which in the future will be placed in tbtc-v2 SDK)
for integrating with contracts, and create a context for ledger live app where
we will keep connected ethereum and bitcoin addresses.

The LedgerLiveAppEthereumSigner will use `@ledgerhq/wallet-api-client` package,
and for the rest things (connecting ethereum wallet, connecting bitcoin wallet,
sending bitcoins to address), we will use `wallet-connect-client-react` lib.

The first step is to install the `wallet-connect-client-react`