Rewamp (3) - LedgerLiveAppContext
This context is where we will store our eth and btc addresses for ledger live
app.

Why we need it?

We have to detect when the address is changed in our LedgerLiveApp and then
update our signer inside threshold lib in `ThresholdContext`. For website we
could just detect it with our `useWeb3React` hook, but in this case it is not
that simple. The only way in the previous implementation was to get the actual
address from the redux store, but this required to move `<ReduxProvider>` above
the `<ThresholdProvider`, which might broke the app in some places.

That's why I've decided to create a separate context foe Ledger Live App and
keep all data about connected accounts there. We will put the context above
`ThresholdContext` so we can use it there and later we will make sure to save
the account there each time we use `useRequestAccount` hook from
`wallet-api-client-react`.