Rewamp (4) - use ledger's wallet api for react
This is probably final step (or one of the final steps) in this rewamp. There
are few bigger changes in this commit:

1. Remove `LedgerLiveAppManager` completely

We won't need that class anymore. The only thing we need is
`LedgerLiveAppEthereumSigner` to interact with eth contracts. The rest thigns
(like connecting wallets, and sending bitcoins) will be handled with hooks from
ledger's wallet-api for react.

2. Store `LedgerLiveAppEthereumSigner` instance in our `TBTC` class in
`thershold-ts` lib.

We also remove the need to pass anything related to ledger live app through
`ThresholdConfig`, like we did with `LedgerLiveAppManager`. The Ledger Live App
Signer will be created inside the class if necessary.

3. Refactor `useRequestEthereumAccount` and `useRequestBitcoinAccount`

Those hooks will now use `useRequestAccount` hook from
`@ledgerht/wallet-api-client-react`, but additionaly the `requestAccount` method
will save the bitcoin/ethereum account in the LedgerLiveContext and ethereum
account in the LedgerLiveAppEthereumSigner (through TBTC method).