Refactor useEffect inside useRequestAccount hooks
For both bitcoin and ethereum hooks we can just save the account address
immidietely to the LedgerLiveApp Context - we don't have to check if the address
is undefined or not in the if statements like before.

Additionally we can also save the eth account for the
LedgerLiveAppEthereumSigner using one-liner.

I've also remove saving bitcoin address to the LedgerLiveApp signer. It was a
mistake. Only eth addresses should be saved there, since it is a signer for
ethereum chain.