Pass ledgerLiveAppEhereumSigner to Threshold lib
I've decided to add additional property to thresholdConfig.ethereum:
`ledgerLiveAppEthereumSigner`. This will allow us to pass the signer when
creating a new instance of TBTC.

It will be useful when connecting account. Our threshold library reinitializes
all classed inside and we don't want to lose the data from our signer
(especcially the connected address that we store there). That's why we pass an
old signer from therhols lib to `updateConfig` method when we connect an
account.