Implement Resume Deposit with SDK v2
Implements resuming deposit (from file) with SDK v2.

For that we have to create deposit receipt from the values that we store in the
file. Since those values are in plain text we have to change them to `Hex`
objects (except `depositor`). Then, we initiate deposit object from receipt and
calculate the correct btc deposit address.