Implement reveal deposit for ledger live app
Implements reveal deposit functionalityin ledger live app. For this we create a
separate `useSendLedgerLiveAppTransactionFromFn` hook, that is a copy of
`useSendTransactionFromFn`, but changed in a way to work with our LedgerLiveApp
Signer.

We then check which hook should we use inside `useRevealDepositTrnasaction`,
based on the `isEmbed` flag.

We are also adding `await` to `sendTransaction` method in
`LedgerLiveAppEthereumSigner`, because we need to return that transaction
response (not promise) from that function.