Fix fetching Bridge Activity in Ledger Live App
The fetching was not working in Ledger Live App, because it used the account
from `useWeb3React` hook, which doesn't work here. I've replaced it with
`useIsActive` hook, which returns either web3 account or connected ledger
account based on `isEmbed` flag.

The bridge activity still uses the old version of tbtc-v2.ts, but I don't think
it's necessary to fix that in this PR. The main reason to use a new SDK in this
PR is to see if the LedgerLiveAppEthereumSigner will work properly with it.