Fix `no ongoingRequest` error
There was an error when doing a reveal that displayed `no ongoingReques:
<request_id>`. It was happening because we were using a separate transport
object in our LedgerLiveAppEthereumSigner, and a separate one for our react
hooks. The one in our react hook was always connected and we didn't disconnect
it after doing a transaction. This way, when we used our LedgerLiveApp Signer we
had two transport objects connected, and two returned the same message, but the
second time the message was already cleared.

As a workaround we will save the transport object that we use in react in a
separate context and we connect it only before doing a transaction, and
disconnect it right after.