Fix fetching token amount in Ledger Live App
For the tbtc balance to fetch property (in Ledger Live App) we have to use the
`useIsActive` hook to get the account based on the `isEmbed` flag. This will
return Leger Live App connected account, if the app is run in Ledger Live, or
web3 account, if the app is run on the browser.