Implement unminting flow with SDK v1
Note: This flow will work only with the newest version of tbtc-v2.ts that is not
yet released. Once it is I will update the version in a separate commit.

Implements whole unminting process with tbtc-v2.ts SDK (v2). The unminting process
is being implemented in a separate PR, so this is a quick implementation to make
the whole process easier once the implementation PR is merged (and to make sure
it will work with new SDK).

For this, I've created a separate method in our `TBTC` class:
`requestRedemptionSdkV2`, which is then used in `useRquestRedemption` hook.