Add a posisbility to send exact amount of BTC
Adds an input form that allows the user to enter bitcoin amount that he wants to
send to the deposit address. It will only show when `isEmbed` flag is set to
true.

There were some additional changes required to make it possible.

First of all, I've adedd additional parameter to `TokenBalanceInputProps` -
`tokenDecimals`. This way it will be possible to pass that parameter to
`TokenBalanceInput` component through `FormikTokenBalanceInput` component, and
it will allow to specify the input for the token we want to use. For example,
without that, whenever we put the amount in `FormikTokenBalanceInput`, it will
automatically convert that to wei (amount*10^18), which means that we can't use
it with bitcoin, that has only 8 decimal places. After the change, we can pass
`tokenDecimals = 8`, to make the `FormikTokenBalanceInput` working for bitcoin
amounts.

Second of all, there were some changes needed in `stc/utils/forms`, espcially
with the `defaultGreaterThanMsg` and `defaultLessThanMsg`. The form uses
`formatTokenAmount` which also worked with default 18 decimals and 2 precision.
This means, that when we've put a value that is greater than our bitcoin balance
(let's say our btc balance is 0.66666), it would siplay an error: The value
should be less than or equal `~0.67` + it would break the app. To fix that we
have to pass 8 decimals nad 8 precision there.