Move types back to threshold-ts/tbtc/index.ts
We should keep our types as close to the related class as possible, moving it
out of the tbtc directory in our threshold-ts is not a good idea, since it's
closely related to that "module". For now, to not change too much code, let's
change it back to what it was to keep things consistent. In every module in
Threshold lib we keep the types in the index file.

We might think about separating the file to smaller pieces/classes like one
related to deposit and one related to redemption but, this is a work for another
PR.