Add new method: initiate a deposit from receipt
Adds new method in our TBTC class: `initiateDepositFromReceipt`. This method
allow us to initiate deposit object straight from redux store, and this is what
we do in our findUtxo effect. This fixes an error that occured when user
generated a deposit address, disconnected an account at step 2, and then
reconnect. WIthout this commit there will be an infinite loading.

Besides that I had to remove the `if (!this_deposit)` condition from the deposit
getter, because it makes it impossible to check if the deposit is empty without
throwing and error. We could create a separate method that checks that but I
don't think it's necessary.