Clean out TBTC class a little
Clean out TBTC class a little (in threshold-ts lib).The main thing here was to
not use artifacts from `@keep-network/tbtc-v2" lib and we made it! It now
imports mainnet and goerli artifacts from the new SDK v2 and we can get it by
calling `getTbtcV2Artifact`.

Beisdes that we are know saving ethereum config in a privae property of TBTC
class (just like we do with bitcoinConfig). This allow us to get ethereum chain
id in any of the function and we need that in some places to get an artifact
(based on the chain id), and the get the block number that it was depoloyed to.
We use that block number to look for the specific events.