Fix unminting process
There was a bug where after doing an unmint the dApp did not redirect to the
UnmintingDetails page. It behaved like that because it wrongly builded a
redemption key - we passed already prefixed `walletPublicKey` and
`redeemerOutputScript`, and the it was prefixed again in sdk lib. I've fixed it
by extracting only the code that was necessary from that funcion (so I'm acually
reverting the changes).

Besides that I've made it possible to pass additional argument to success
function in `useSendTransactionFromFn`. It was neede because
`findWalletForRedemption` is not a publi function anymore in our SDK and the
wallet is found during a redemption request (and the method from SDK return tx
hash of the redemption request and wallet public key hash). We return both of
them in our TBTC class (in threshold-ts lib) and the pass it ot `onSuccess`
callback in the hook I mentioned earlier. This way we could also get rid of
`findWalletForRedemption` method and all things related to that.

I've also managed to remove unnecessary imports. It looks like there are no
import from tbtc-v2 or tbtc-v2.ts in our threshold-ts/tbtc/index.ts file (at
least not direct ones :D).