Reduce the use of Hex class
To simplify the `Hex to string conversion` while processing transactions
in the dApp there was made a decision to reduce the frequency of Hex
class usage and use it only in the scope of the `threshold-lib` (and its
utility functions). It reduces the complexity of the code and parameters
transformation on the callstack. Data records stored in the global store
are required to have serializable form (eg. strings) so now it's easier
to spot the transformation end and analyze transactions processing.