Refactor disabled button for token amount form
The previosu commit introduces `shouldDisableButton` property which might get
a little confusing, because when it's set to true someone might expect that the
button is disabled no matter what, when in reality it relies on `isDisabled`
property also.

Instead of this, I've added `isDisabled` property for `Connect Wallet` button in
SubmitTxButton component and set it to false. This  overrides the `isDisabled`
property from `...buttonProp`, because in this case we don't want ot have this
button disabled anyway.

This means that we can use `isDisabled` to disable token amount form if the
account is not connected but keep the `Connect Wallet` button active.