Close modal after successful deposit reveal.
We've redesigned the step 3 in 5d1aa5544cab54bedd4db565fae97289d31d4028, where
we moved most of the logic from `TbtcMintingConfirmation` modal to
`InitiateMinting` component, but we did not include `closeModal()` call that was
in `onSuccessfulDepositReveal` method. This is why the `TransactionIsPending`
modal was not closed when the deposit was revealed.