Exclude BigNumber from useEffect's dependency
We've used value from utxo as a dependency in our useEffect, which was a type of
a `BigNumber`. This will produce unencessary re-renders, because useEffect uses
shallow equality comparison to identify whether dependencies are updated so when
it compares (for example) `BigNumber.from("5")` with `BigNumber.from("5")` it
will return false.

Changing the value to string first will resolve that issue.

Besides that I've also removed the Skeleton for btc value, because the utxo is
always defined so it should have value by the time we are at this stage.