Improve button tracking for minting flow
Add additional parameters to generate deposit address button and initiate
minting address button. For both of these autocapture events we are adding two
additional properties:
- button-id - id of the button that should not be changed
- button-text - text that is placed inside of the button.
For both of these, we still keep the `button-name` property if we want backward
compatibility with the data that was gathered until this point.

For initiate minting button we've also added `deposited-btc-amount` property
which is basicaly a btc amount (in satoshi) that was extracted from the utxo
detected from deposit address in the current minting flow.