Removes `loading` from dependency array
Remove `loading` from dependency array in `useCaptureWalletConnected` hook.
Putting loading there might lead to unexpected `Wallet Connected` event sent,
because `loading` can change it's value several times when the account is
connected. It will be enough if we keep `isLoadedFromConnectedAccount` as it
will be changed only one time for the current account.